requirements.txt
fastapi
uvicorn
groq
jsonref
python-dotenv
httpx
pydantic
anyio